separate test for variable argument middlewares
