+++ test/express-style.js
@@ -51,24 +51,4 @@ describe('express style middleware', function () {
       }
     })
   })
-
-  it('combine multiple argument middlewares', function () {
-    const m1 = (a, b, results, next) => {
-      results.push('m1 called')
-      next()
-    }
-    const m2 = (a, b, results, next) => {
-      results.push('m2 called')
-      next()
-    }
-    const m3 = (a, b, results, next) => {
-      results.push('m3 called')
-      next()
-    }
-
-    const combined = combineMiddlewares(m1, m2, m3)
-    const results = []
-    combined(null, null, results)
-    assert.deepEqual(results, ['m1 called', 'm2 called', 'm3 called'])
-  })
 })
